data_handling/db_handle.py - db_name is now an instance variable
db_name no longer needs to be passed for every db_handle function.
The intent is that an instance of the db_handle class will correspond
to a connection to a database, so the name only needs to be given when
first connecting to the database.  A new instance of db_handle should
be created for connections to a difference sql database.