main.py - Finished Minimum Viable Functionality for Main.py Trading Heart
This seems to be working as expected, however, there is no longer any action
on the public sandbox, so it is hard to tell.

As of now, The trading algorithm is to place a buy order at the same price as
the getTicker() funciton from GDAX.  A sell order is then placed slightly higher
than the price it was bought at.  The quantity and amount higher can be programmed
in bot_params.json via the quantity and aggression parameters, respectively.