main.py - Revise basic trading logic.  Currently does not work because of unreliability of the buy() and sell() APIs
The trading heart will simply open a buy order at a price above the
ticker price (adjustable in bot_params).  It will then immediately
place a sell order for a price of  a certain factor of the ticker price
(also adjustable in bot_params).  This whole process is repeated only if there
no open orders.